refactor(web): remove redux, add recoil
This swaps state management framework from [`redux`](https://redux.js.org/) to [`recoil`](https://recoiljs.org/).

Recoil is more modern, ergonomic, plays nicely with React hooks, and, hopefully with async rendering.

I am planing to introduce some serious modifications to the app state shortly, so it is better to make the swap before that, to avoid basically writing the thing twice.

So far we used redux for just a couple of things:
- the state of the "Tooltip sort by" dropdown and checkbox
- getting current page pathname in the navbar component (via connected-next-router)

Both are kind of redundant also. I made navbar to use the pathname from the Next.js router, and moved dropdown state into a recoil atom.

In this PR, there should be no visual or behavioral changes in the application. If there is, it's a bug.