fix: don't sort clusters for "per country" plots
This removes sorting of cluster state for "per country" page introduced in https://github.com/hodcroftlab/covariants/pull/279
This ensures the old ordering of colored areas on area plots.

A complication of that removal is that sidebar component requires sorted clusters. However without it being sorted globally in case of "per country" page, it comes unsorted into the component. So sidebar now additionally performs the sorting.