fix(web): crash when accessing "mutations" pages on side bar
This was due to the mutations counts missing for these pages. It used to be displaying a placeholder, but at some point I migrated to async data fetching and it was throwing an exception which isn't handled correctly.

I am now explicitly ignoring the error if the mutation counts json is missing for a variant - and in this case nothing is shown on the page where the mutation counts widget should be.